Make keycode mapping lazy
This should allow other keycodes, outside value 255, to be found at runtime. It should also allow
AXElements to start up a bit faster since it won't perform the lookups until needed.

This doesn't actually work yet because it requires more complex logic for the lazy loading. Lookups are
done from keycode to character with UCKeyTranslate(), but we want to do the opposite, so we have to
do something like load characters in chunks when there is a cache miss and then error out if we get to
a certain point.